Add houses wip

+++ core/src/com/netrew/game/pathfinding/FlatTiledGraph.kt
@@ -111,7 +111,7 @@ class FlatTiledGraph : TiledGraph<FlatTiledNode?> {
         return node!!.index
     }
 
-    fun getNodeByPosition(position: Vector2, tileSize: Int): FlatTiledNode {
+    fun getNodeByPosition(position: Vector2, tileSize: Float): FlatTiledNode {
         val x = (position.x / tileSize).toInt()
         val y = (position.y / tileSize).toInt()
         return getNode(x, y)
